refactor(rust): Extract tests to separate files to fix line count CI
Move tests from pubsub.rs, router.rs, and scheduler.rs into separate
test modules following the existing crate pattern (e.g., client/tests.rs).
This reduces line counts to under 1000 lines per file:
- pubsub.rs: 1366 -> 990 lines
- router.rs: 1009 -> 787 lines
- scheduler.rs: 1055 -> 786 lines

Also condensed some documentation comments to further reduce line counts.
